Validation of Hex files and Clean-up
