+++ Calliope App/ViewsAndControllers/EditorsAndPrograms/en.lproj/EditorAndPrograms.strings
@@ -57,7 +57,7 @@
 "bzf-ki-nbR.normalTitle" = "Choose File";
 
 /* No comment provided by engineer. */
-"5dO-ZD-bhg.text" = "Open programs to save them or transfer programs directly to your Calliope mini.";
+"5dO-ZD-bhg.text" = "Transfer programs from your iPhone or iPad directly to your Calliope mini or save them in your Calliope mini folder.";
 
 /* No comment provided by engineer. */
-"V6G-YI-Xec.text" = "Saved programs in your Calliope mini folder can be opened, transferred and shared.";
+"V6G-YI-Xec.text" = "A long press takes you to the option of transferring, sharing and deleting your programs stored in the Calliope mini folder.";
